requirements.txt
asgiref==3.4.1
astroid==2.6.2
autopep8==1.5.7
boto3==1.17.111
botocore==1.20.111
certifi==2021.5.30
charset-normalizer==2.0.1
colorama==0.4.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.5
django-amazon-ses==4.0.0
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-hosts==5.1
django-import-export==2.7.1
django-jazzmin==2.4.9
django-phonenumber-field==6.0.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.42
Pillow==9.0.0
psycopg2-binary==2.9.3
pycodestyle==2.7.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-whois==0.7.3
pytz==2021.1
PyYAML==6.0
requests==2.27.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
stripe==2.65.0
tablib==3.2.0
toml==0.10.2
urllib3==1.26.6
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
django-heroku==0.3.1
